% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{BIC.cox_cure}
\alias{BIC.cox_cure}
\title{Bayesian Information Criterion (BIC)}
\usage{
\method{BIC}{cox_cure}(object, ...)
}
\arguments{
\item{object}{An object for a fitted model.}

\item{...}{Other objects.}
}
\description{
Compute Bayesian information criterion (BIC) or Schwarz's Bayesian criterion
(SBC) for possibly one or several objects.
}
\examples{
## See examples of function 'cox_cure'.
}
