% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{BIC.cox_cure_net}
\alias{BIC.cox_cure_net}
\title{Bayesian Information Criterion (BIC)}
\usage{
\method{BIC}{cox_cure_net}(object, ...)
}
\arguments{
\item{object}{An object for a fitted model.}

\item{...}{Other arguments for future usage.}
}
\description{
Compute Bayesian information criterion (BIC) or Schwarz's Bayesian criterion
(SBC) from a fitted model.
}
\examples{
## See examples of function 'cox_cure_net'.
}
