% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{BIC.cox_cure_uncer}
\alias{BIC.cox_cure_uncer}
\title{Bayesian Information Criterion (BIC)}
\usage{
\method{BIC}{cox_cure_uncer}(object, ..., method = c("obs",
  "certain-event"))
}
\arguments{
\item{object}{An object for a fitted model.}

\item{...}{Other objects.}

\item{method}{A character string specifying the method for computing the BIC
values.  The available options are \code{"obs"} for regular BIC based on
the number of observations, and \code{"certain-event"} for a variant of
BIC based on the number of certain uncensored events.  The former is
used by default.}
}
\description{
Compute Bayesian information criterion (BIC) or Schwarz's Bayesian criterion
(SBC) for possibly one or several objects.
}
\examples{
## See examples of function 'cox_cure'.
}
