% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{cIndex}
\alias{cIndex}
\title{Weighted Concordance Index}
\usage{
cIndex(time, event, risk_score, weight = NULL)
}
\arguments{
\item{time}{A numeric vector for observed times}

\item{event}{A numeric vector for event indicators.}

\item{risk_score}{A numeric vector for risk scores.}

\item{weight}{A optional numeric vector for weights.  By default, equal
weights are used.}
}
\description{
This function is a straightforward implementation of concordance index
(C-index or C-statistic) that allows weights.  Asano and Hirakawa (2017)
proposed using the probability of being susceptible as weights for cure
models.  It reduces to Harrell's C-index for equal weights.
}
\details{
Let \eqn{r_i}, \eqn{t_i}, and \eqn{\delta_i} denote the risk score, observed
time, and event indicator of \eqn{i}-th subject.  The pair of
\eqn{(t_i,\delta_i)} and \eqn{(t_j,\delta_j)}, where \eqn{i<j}, are defined
to be comparable if \eqn{t_i<t_j,\delta_i=1} or
\eqn{t_i=t_j,\delta_i=1,\delta_j=0}.  In the context of survival analysis,
the risk scores of a comparable pair are said to be concordant with the
survival outcomes if \eqn{r_i>r_j}.  The C-index is defined as the
proportion of the concordant pairs among the comparable pairs.  For
comparable pair satisfying \eqn{t_i<t_j,\delta_i=1}, we count 0.5 in the
numerator of the concordance index for tied risk scores (\eqn{r_i=r_j}).
}
\references{
Asano, J., & Hirakawa, A. (2017). Assessing the prediction accuracy of a
cure model for censored survival data with long-term survivors: Application
to breast cancer data. Journal of biopharmaceutical statistics, 27(6),
918--932.

Harrell, F. E., Lee, K. L., & Mark, D. B. (1996). Multivariable prognostic
models: Issues in developing models, evaluating assumptions and adequacy,
and measuring and reducing errors. Statistics in medicine, 15(4), 361--387.
}
