% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\docType{methods}
\name{coef,coxphx-method}
\alias{coef,coxphx-method}
\title{Estimated Coefficients of Covariates}
\usage{
\S4method{coef}{coxphx}(object, ...)
}
\arguments{
\item{object}{\code{\link{coxphx-class}} object.}

\item{...}{Other arguments for future usage.}
}
\value{
A named numeric vector.
}
\description{
\code{coef,coxphx-method} is an S4 class method that extracts covariate
coefficient estimates from \code{\link{coxphx-class}} object from
function \code{\link{coxphx}}.
}
\examples{
## See examples given in function coxphx.
}
\seealso{
\code{\link{coxphx}} for fitting integrative Cox model;
\code{\link{summary,coxphx-method}} for summary of a fitted model.
}
