% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\docType{methods}
\name{coef,iCoxph-method}
\alias{coef,iCoxph-method}
\title{Estimated Coefficients of Covariates}
\usage{
\S4method{coef}{iCoxph}(object, ...)
}
\arguments{
\item{object}{\code{\link{iCoxph-class}} object.}

\item{...}{Other arguments for future usage.}
}
\value{
A named numeric vector.
}
\description{
\code{coef,iCoxph-method} is an S4 class method that extracts covariate
coefficient estimates from \code{\link{iCoxph-class}} object from
function \code{\link{iCoxph}}.
}
\examples{
## See examples given in function iCoxph.
}
\seealso{
\code{\link{iCoxph}} for fitting integrative Cox model;
\code{\link{summary,iCoxph-method}} for summary of a fitted model.
}
