% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.cox_cure}
\alias{coef.cox_cure}
\title{Estimated Covariate Coefficients}
\usage{
\method{coef}{cox_cure}(object, ...)
}
\arguments{
\item{object}{Object representing a fitted model.}

\item{...}{Other arguments for future usage.}
}
\value{
A named numeric vector.
}
\description{
Extract the covariate coefficient estimates from a fitted Cox cure rate
model.
}
