% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.cox_cure_net}
\alias{coef.cox_cure_net}
\title{Estimated Covariate Coefficients}
\usage{
\method{coef}{cox_cure_net}(object, selection = c("bic", "none"), ...)
}
\arguments{
\item{object}{Object representing a fitted model.}

\item{selection}{A character string for specifying the criterion for
selection of coefficient estimates.  The available options are
\code{"bic"} for selecting coefficient estimates based on BIC criterion
and \code{"none"} for not performing selection and returning the whole
solution path.}

\item{...}{Other arguments for future usage.}
}
\value{
A named numeric vector.
}
\description{
Extract the covariate coefficient estimates from a fitted Cox cure rate
model.
}
