% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_cure_mar.R
\name{cox_cure_mar}
\alias{cox_cure_mar}
\alias{cox_cure_mar.fit}
\alias{cox_cure_net_mar}
\alias{cox_cure_net_mar.fit}
\title{Cox Cure Rate Model For Uncertain Event Indicators}
\usage{
cox_cure_mar(
  surv_formula,
  cure_formula,
  mar_formula,
  time,
  event,
  data,
  subset,
  contrasts = NULL,
  bootstrap = 0,
  control = cox_cure.control(),
  surv_control = intsurv.control(),
  cure_control = intsurv.control(),
  mar_control = intsurv.control(),
  ...
)

cox_cure_mar.fit(
  surv_x,
  cure_x,
  mar_x,
  time,
  event,
  cure_intercept = TRUE,
  mar_intercept = TRUE,
  bootstrap = 0,
  control = cox_cure.control(),
  surv_control = intsurv.control(),
  cure_control = intsurv.control(),
  mar_control = intsurv.control(),
  ...
)

cox_cure_net_mar(
  surv_formula,
  cure_formula,
  mar_formula,
  time,
  event,
  data,
  subset,
  contrasts = NULL,
  cv_nfolds = 0,
  control = cox_cure.control(),
  surv_control = cox_cure_net.control(),
  cure_control = cox_cure_net.control(),
  mar_control = cox_cure_net.control(),
  ...
)

cox_cure_net_mar.fit(
  surv_x,
  cure_x,
  mar_x,
  time,
  event,
  cure_intercept = TRUE,
  mar_intercept = TRUE,
  cv_nfolds = 0,
  control = cox_cure.control(),
  surv_control = cox_cure_net.control(),
  cure_control = cox_cure_net.control(),
  mar_control = cox_cure_net.control(),
  ...
)
}
\arguments{
\item{surv_formula}{A formula object starting with \code{~} for the model
formula in survival model part.  For Cox model, no intercept term is
included even if an intercept is specified or implied in the model
formula.  A model formula with an intercept term only is not allowed.}

\item{cure_formula}{A formula object starting with \code{~} for the model
formula in incidence model part.  For logistic model, an intercept term
is included by default and can be excluded by adding \code{+ 0} or
\code{- 1} to the model formula.}

\item{mar_formula}{A formula specifying a logistic model for modeling the
missingness of the event indicators under the missing-at-random (MAR)
assumption.}

\item{time}{A numeric vector for the observed survival times.}

\item{event}{A numeric vector representing the event indicators, where
\code{NA}'s are allowed and represent uncertain event status.}

\item{data}{An optional data frame, list, or environment that contains the
model covariates and response variables (\code{time} and \code{event})
If they are not found in data, the variables are taken from the
environment of the specified formula, usually the environment from which
this function is called.}

\item{subset}{An optional logical vector specifying a subset of observations
to be used in the fitting process.}

\item{contrasts}{An optional list, whose entries are values (numeric
matrices or character strings naming functions) to be used as
replacement values for the contrasts replacement function and whose
names are the names of columns of data containing factors.  See
\code{contrasts.arg} of \code{\link[stats]{model.matrix.default}} for
details.}

\item{bootstrap}{An integer representing the number of bootstrap samples for
estimating standard errors of the coefficient estimates.  The bootstrap
procedure will not run if \code{bootstrap = 0} by default.  If
\code{bootstrap > 0}, the specified number of bootstrap samples will be
used for estimating the standard errors.}

\item{control}{A \code{cox_cure.control} object that contains the control
parameters.}

\item{surv_control}{A list of control parameters for the survival model
part.  One may use the help functions \code{intsurv.control()} or
\code{cox_cure_net.control} to specify the available options.}

\item{cure_control}{A list of control parameters for the logistic model
part similar to \code{surv_control}.}

\item{mar_control}{A list of control parameters for the MAR model part.}

\item{...}{Other arguments passed to the control functions for backward
compatibility.}

\item{surv_x}{A numeric matrix for the design matrix of the survival model
component.}

\item{cure_x}{A numeric matrix for the design matrix of the cure rate model
component.  The design matrix should exclude an intercept term unless we
want to fit a model only including the intercept term.  In that case, we
need further set \code{cure_intercept = FALSE} to not standardize the
intercept term.}

\item{mar_x}{A numeric matrix representing the design matrices for the MAR
model part.}

\item{cure_intercept}{A logical value specifying whether to add an intercept
term to the cure rate model component.  If \code{TRUE} by default, an
intercept term is included.}

\item{mar_intercept}{A logical value specifying whether to add an intercept
term to the MAR model part.  If \code{TRUE} by default, an intercept
term is included.}
}
\value{
A \code{cox_cure_mar} object for the fitted Cox cure rate model with
    uncertain event indicators.
}
\description{
For right-censored data with uncertain event indicators, fit the Cox cure
rate model proposed by Wang et al. (2020).
}
\references{
Wang, W., Luo, C., Aseltine, R. H., Wang, F., Yan, J., & Chen, K. (2020).
Suicide Risk Modeling with Uncertain Diagnostic Records. \emph{arXiv
preprint arXiv:2009.02597}.
}
