% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_cure.R
\name{cox_cure_net}
\alias{cox_cure_net}
\title{Regularized Cox Cure Rate Model}
\usage{
cox_cure_net(
  surv_formula,
  cure_formula,
  time,
  event,
  data,
  subset,
  contrasts = NULL,
  cv_nfolds = 0,
  control = cox_cure.control(),
  surv_control = cox_cure_net.control(),
  cure_control = cox_cure_net.control(),
  ...
)
}
\arguments{
\item{surv_formula}{A formula object starting with \code{~} for the model
formula in survival model part.  For Cox model, no intercept term is
included even if an intercept is specified or implied in the model
formula.  A model formula with an intercept term only is not allowed.}

\item{cure_formula}{A formula object starting with \code{~} for the model
formula in incidence model part.  For logistic model, an intercept term
is included by default and can be excluded by adding \code{+ 0} or
\code{- 1} to the model formula.}

\item{time}{A numeric vector for the observed survival times.}

\item{event}{A numeric vector for the event indicators.}

\item{data}{An optional data frame, list, or environment that contains the
model covariates and response variables (\code{time} and \code{event})
If they are not found in data, the variables are taken from the
environment of the specified formula, usually the environment from which
this function is called.}

\item{subset}{An optional logical vector specifying a subset of observations
to be used in the fitting process.}

\item{contrasts}{An optional list, whose entries are values (numeric
matrices or character strings naming functions) to be used as
replacement values for the contrasts replacement function and whose
names are the names of columns of data containing factors.  See
\code{contrasts.arg} of \code{\link[stats]{model.matrix.default}} for
details.}

\item{cv_nfolds}{A nonnegative integer representing the number of folds in
cross-validation.}

\item{control}{A \code{cox_cure.control} object that contains the control
parameters.}

\item{surv_control}{A list of control parameters for the survival model
part.  One may use the help functions \code{intsurv.control()} or
\code{cox_cure_net.control} to specify the available options.}

\item{cure_control}{A list of control parameters for the logistic model
part similar to \code{surv_control}.}

\item{...}{Other arguments passed to the control functions for backward
compatibility.}
}
\value{
\code{cox_cure_net} object for regular Cox cure rate model or
\code{cox_cure_net_mar} object for Cox cure rate model with uncertain
events.
}
\description{
For right-censored data, fit a regularized Cox cure rate model through
elastic-net penalty following Masud et al. (2018), and Zou and Hastie
(2005).  For right-censored data with uncertain event status,
fit the regularized Cox cure model proposed by Wang et al. (2020).  Without
regularization, the model reduces to the regular Cox cure rate model (Kuk
and Chen, 1992; Sy and Taylor, 2000)
}
\details{
The model estimation procedure follows expectation maximization (EM)
algorithm.  Variable selection procedure through regularization by elastic
net penalty is developed based on cyclic coordinate descent and
majorization-minimization (MM) algorithm.
}
\references{

}
