% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fitcox.int.pen}
\alias{fitcox.int.pen}
\title{Penalized integrative estimation of the heterogeneous treatment
effect based on the covariates from the basis approximation}
\usage{
fitcox.int.pen(
  tau.mat,
  c1.sieve,
  c0.sieve,
  lambda.sieve,
  a,
  s,
  t,
  delta,
  penalty.type = "adaptive.lasso",
  nfolds = 5
)
}
\arguments{
\item{tau.mat}{the matrix that contains the covariate terms of \eqn{\tau(X)}.}

\item{c1.sieve}{the covariate matrix for \eqn{c_1(X)} in the sieve
approximation.}

\item{c0.sieve}{the covariate matrix for \eqn{c_0(X)} in the sieve
approximation.}

\item{lambda.sieve}{the covariate matrix for \eqn{\lambda(X)} in the sieve
approximation.}

\item{a}{the binary treatment assignment, where the active treatment group
should be encoded as \code{1}, and the control group should be encoded as \code{0}.}

\item{s}{the binary data source indicator, where the trial data should be
encoded as \code{1}, and the observational study should be encoded as \code{0}.}

\item{t}{the observed event time.}

\item{delta}{the event indicator.}

\item{penalty.type}{the type of penalty in the variable selection procedure,
including
\itemize{
\item \code{lasso}: the lasso penalty.
\item \code{adaptive.lasso}: the adaptive lasso penalty.
}}

\item{nfolds}{the fold of cross-validation, the default value is \code{5}.}
}
\value{
a list of estimators, including
\itemize{
\item \code{betahat}: the estimated treatment effect parameter
\item \code{ve.beta}: the variance estimate of the estimated treatment effect
\item \code{var.betahat}: the covariance matrix of the estimated treatment effect
parameter
\item \code{ate}: the estimated average treatment effect
\item \code{ve.ate}: the variance estimate of the average treatment effect
\item \code{psi.hat}: the estimated confounding function parameter
}
}
\description{
This function obtains a penalized integrative estimator with its
variance estimates by minimizing the penalized log partial likelihood.
The penalty function can selected from the adaptive lasso and the
lasso.
}
