% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{iCoxph-class}
\alias{iCoxph-class}
\title{An S4 Class to Represent a Fitted Extended Cox Model}
\description{
\code{iCoxph-class} is an S4 class that represents a fitted model.
Function \code{\link{iCoxph}} produces objects of this class.
See ``Slots'' for details.
}
\section{Slots}{

\describe{
\item{\code{call}}{Function call.}

\item{\code{formula}}{Formula.}

\item{\code{nObs}}{A positive integer.}

\item{\code{data}}{A data frame.}

\item{\code{estimates}}{A list.}

\item{\code{control}}{A list.}

\item{\code{start}}{A list.}

\item{\code{na.action}}{A length-one character vector.}

\item{\code{xlevels}}{A list.}

\item{\code{contrasts}}{A list.}

\item{\code{convergCode}}{A nonnegative integer.}

\item{\code{partLogL}}{A numeric value.}

\item{\code{logL}}{A numeric value.}
}}

\seealso{
\code{\link{iCoxph}} for details of slots.
}
