% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{iCoxph.summary-class}
\alias{iCoxph.summary-class}
\title{An S4 Class to Represent Summary of a Fitted Model}
\description{
\code{iCoxph.summary-class} is an S4 class that represents a summarized
model.  Function \code{\link{summary,iCoxph-method}} produces objects of
this class.  See ``Slots'' for details.
}
\section{Slots}{

\describe{
\item{\code{call}}{Function call.}

\item{\code{coefMat}}{A matrix.}

\item{\code{logL}}{A numeric value.}
}}

\seealso{
\code{\link{summary,iCoxph-method}} for meaning of slots.
}
