% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsurv.R
\docType{package}
\name{intsurv-package}
\alias{intsurv-package}
\title{Integrative Survival Modeling}
\description{
The package \pkg{intsurv} mainly provides function fitting extended Cox
proportional hazard model for uncertain survival data due to imperfect data
integration.
}
\references{
Wang, W., Chen, K., & Yan, J. (2017+).  Extended Cox Model by ECM Algorithm
for Uncertain Survival Records Due to Imperfect Data Integration. (working
in progress)
}
