% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sieve.mat}
\alias{sieve.mat}
\title{Generate the matrices of basis functions by the method of sieves for
all nuisance functions}
\usage{
sieve.mat(
  c1.mat,
  c1.discrete = NULL,
  q.c1,
  c0.mat,
  c0.discrete = NULL,
  q.c0,
  lambda.mat,
  lambda.discrete = NULL,
  q.lambda,
  basis.type = "poly.raw"
)
}
\arguments{
\item{c1.mat}{the matrix that contains the continuous covariates which may
affect \eqn{c_1(X)}.}

\item{c1.discrete}{the matrix that contains the discrete covariates which may
affect \eqn{c_1(X)}. The default is \code{NULL}.}

\item{q.c1}{the number of basis functions to approximate each continuous
covariate in \eqn{c_1(X)}. Here, it corresponds to the degree of freedom to
generate the polynomial basis.}

\item{c0.mat}{the matrix that contains the continuous covariates which may
affect \eqn{c_0(X)}.}

\item{c0.discrete}{the matrix that contains the discrete covariates which may
affect \eqn{c_0(X)}. The default is \code{NULL}.}

\item{q.c0}{the number of basis functions to approximate each continuous
covariate in \eqn{c_0(X)}. Here, it corresponds to the degree of freedom to
generate the polynomial basis.}

\item{lambda.mat}{the matrix that contains the continuous covariates which
may affect \eqn{\lambda(X)}.}

\item{lambda.discrete}{the matrix that contains the discrete covariates
which may affect \eqn{\lambda(X)}.}

\item{q.lambda}{the number of basis functions to approximate each continuous
covariate in \eqn{\lambda(X)}. Here, it corresponds to the degree of freedom
to generate the polynomial basis.}

\item{basis.type}{the type of basis used to approximate the nuisance functions,
including
\itemize{
\item \code{poly.raw}: the raw polynomial basis. The default type of the basis.
\item \code{poly.orthogonal}: the orthogonal polynomial basis
\item \code{bs}: the B-spline basis for a polynomial spline
}}
}
\value{
a list of covariate matrices for the nuisance functions, including
\itemize{
\item \code{c1.sieve}: the covariate matrix for \eqn{c_1(X)} in the sieve
approximation
\item \code{c0.sieve}: the covariate matrix for \eqn{c_0(X)} in the sieve
approximation
\item \code{lambda.sieve}: the covariate matrix for \eqn{\lambda(X)} in the sieve
approximation
}
}
\description{
This function generate polynomial splines that are needed to approximate
the nuisance functions \eqn{c_1(X)}, \eqn{c_0(X)}, and \eqn{\lambda(X)} by
the method of sieves.
}
