% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuData.R
\name{simuWeibull}
\alias{simuWeibull}
\title{Generate Simulated Data from Cox Model}
\usage{
simuWeibull(nSubject = 1000, beta0Vec, xMat, maxNum = 2,
  nRecordProb = c(0.9, 0.1), matchCensor = 0.1, matchEvent = 0.1,
  censorMin = 0.5, censorMax = 12.5, lambda = 0.005, rho = 0.7,
  fakeLambda1 = lambda * exp(-3), fakeRho1 = rho,
  fakeLambda2 = lambda * exp(3), fakeRho2 = rho, mixture = 0.5,
  randomMiss = TRUE, eventOnly = FALSE, ...)
}
\arguments{
\item{nSubject}{Number of subjects}

\item{beta0Vec}{Time-invariant covariate coefficients.}

\item{xMat}{Design matrix. By default, three continuous variables following
standard normal distribution and one binary variable following bernoulli
distribution with equal probability are used.}

\item{maxNum}{Maximum number of uncertain records.}

\item{nRecordProb}{Probability of the number of uncertian records.}

\item{matchCensor}{The matching rate for subjects actually having censoring
times.}

\item{matchEvent}{The matching rate for subjects actually having event
times.}

\item{censorMin}{The lower boundary of the uniform distribution for
generating censoring time.}

\item{censorMax}{The upper boundary of the uniform distribution for
generating censoring time.}

\item{lambda}{A positive number, scale parameter in baseline rate function
for true event times.}

\item{rho}{A positive number, shape parameter in baseline rate function for
true event times.}

\item{fakeLambda1}{A positive number, scale paramenter in baseline rate
function for fake event times from one distribution.}

\item{fakeRho1}{A positive number, shape parameter in baseline rate function
for fake event times from one distribution.}

\item{fakeLambda2}{A positive number, scale paramenter in baseline rate
function for fake event times from another distribution.}

\item{fakeRho2}{A positive number, shape parameter in baseline rate function
for fake event times from another distribution.}

\item{mixture}{The mixture weights, i.e., the probabilies (summing up to
one) of fake event times coming from different mixture components.}

\item{randomMiss}{A boolean value specifying whether the labels of the true
records are missing completely at random (MCAR) or missing not at random
(MNAR). The default value is \code{TRUE} for MCAR.}

\item{eventOnly}{A boolean value specifying whether the uncertain records
only include possible events. The default value is \code{FALSE}, which
considers the censoring cases as the possible truth in addition to event
records.}

\item{...}{Other arguments for future usage.}
}
\value{
A data frame with the following columns, \itemize{ \item \code{ID}:
    subject ID \item \code{obsTime}: observed event times \item
    \code{eventInd}: event indicators \item \code{latentInd}: latent label
    indicating the true recorods } and the corresponding covariates.
}
\description{
This function generates simulated survival data with uncertain records.
The baseline hazard function is parametric Weibull model
and the censoring times follow uniform distribution.
}
\examples{
dat <- simuWeibull(nSubject = 100)
}
\references{
Bender, R., Augustin, T. and Blettner, M. (2005),
Generating survival times to simulate Cox proportional hazards models.
Statistics in Medicine, 24: 1713--1723. doi:10.1002/sim.2059.

Nadarajah, S., & Kotz, S. (2006). R programs for truncated
distributions. Journal of Statistical Software, 16(c02).
}
