% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary,coxphx-method}
\alias{summary,coxphx-method}
\title{Summary of a Fitted Model}
\usage{
\S4method{summary}{coxphx}(object, showCall = TRUE, ...)
}
\arguments{
\item{object}{\code{\link{coxphx-class}} object.}

\item{showCall}{A logic value with default \code{TRUE}, indicating whether
\code{show} method prints out the original call information of
\code{coxphx}.  Set \code{FALSE} for a more concise printout.}

\item{...}{Other arguments for future usage.}
}
\value{
\code{coxphx.summary-class} class object.
}
\description{
For \code{\link{coxphx}} object, the function returns a
\code{\link{coxphx.summary-class}} object whose slots include
\itemize{
    \item \code{call}: Function call of model fitting.
    \item \code{coefMat}: Estimated covariate coefficients.
    \item \code{logL}: Log-likelihood under observed data.
}
}
\examples{
## See examples given in function coxphx
}
\seealso{
\code{\link{coxphx}} for model fitting;
\code{\link{coef,coxphx-method}} for coefficient estimates.
}
