% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,iCoxph-method}
\alias{summary,iCoxph-method}
\alias{summary.iCoxph}
\title{Summary of a Fitted Model}
\usage{
\S4method{summary}{iCoxph}(object, showCall = TRUE, ...)
}
\arguments{
\item{object}{\code{\link{iCoxph-class}} object.}

\item{showCall}{A logic value with default \code{TRUE}, indicating whether
\code{show} method prints out the original call information of
\code{iCoxph}.  Set \code{FALSE} for a more concise printout.}

\item{...}{Other arguments for future usage.}
}
\value{
\code{summary.iCoxph} class object.
}
\description{
For \code{\link{iCoxph}} object, the function returns a
\code{\link{summary.iCoxph}} object whose slots include
\itemize{
    \item \code{call}: Function call of model fitting.
    \item \code{coefMat}: Estimated covariate coefficients.
    \item \code{logL}: Log-likelihood under observed data.
}
}
\examples{
## See examples of function iCoxph
}
\seealso{
\code{\link{iCoxph}} for model fitting;
\code{\link{coef,iCoxph-method}} for coefficient estimates.
}
