% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.cox_cure}
\alias{summary.cox_cure}
\alias{summary.cox_cure_mcar}
\title{Summary of a Fitted Model}
\usage{
\method{summary}{cox_cure}(object, ...)

\method{summary}{cox_cure_mcar}(object, ...)
}
\arguments{
\item{object}{Object representing a fitted model.}

\item{...}{Other arguments for future usage.  A warning will be thrown if
any invalid argument is specified.}
}
\description{
Summarize a fitted Cox cure rate model with possible uncertain event status.
}
