% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_cure_mar.R
\name{cox_cure_mar}
\alias{cox_cure_mar}
\title{For right-censored data with uncertain event status, fit the Cox cure model
proposed by Wang et al. (2020).}
\usage{
cox_cure_mar(
  surv_formula,
  cure_formula,
  mar_forumla,
  time,
  event,
  data,
  subset,
  contrasts = NULL,
  bootstrap = 0,
  control = cox_cure.control(),
  surv_control = intsurv.control(),
  cure_control = intsurv.control(),
  mar_control = intsurv.control(),
  ...
)
}
\arguments{
\item{event}{\code{NA}'s are
allowed and represent uncertain event status.}
}
\value{
A \code{cox_cure_mar} object for the fitted Cox cure rate model with
    uncertain event indicators.
}
\description{
For right-censored data with uncertain event status, fit the Cox cure model
proposed by Wang et al. (2020).
}
\references{
Wang, W., Luo, C., Aseltine, R. H., Wang, F., Yan, J., & Chen, K. (2020).
Suicide Risk Modeling with Uncertain Diagnostic Records. \emph{arXiv
preprint arXiv:2009.02597}.
}
