% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.R
\name{coxph_fit}
\alias{coxph_fit}
\title{Cox proportional hazard model}
\usage{
coxph_fit(time, event, x, start = rep(0, ncol(x)), max_iter = 200,
  rel_tol = 1e-05, early_stop = TRUE, verbose = FALSE)
}
\arguments{
\item{time}{A numeric vector for observed times.}

\item{event}{A numeric vector for event indicators}

\item{x}{A numeric matrix for design matrix.}

\item{start}{A numeric vector for starting values.  By default, an all-zeros
vector with appropriate length (equal to the number of columns of the
design matrix) will be used.}

\item{max_iter}{A positive integer specifying the maximum number of
iterations.}

\item{rel_tol}{A positive number specifying the convergence tolence on the
relative difference based on L1-norm.}

\item{early_stop}{A boolean value specifying whether to stop the iteration
once the negative log-likelihood unexpectedly increases, which may
suggest convergence on likelihood, or indicate numerical issues or
implementation bugs.  The default value is \code{TRUE}.}

\item{verbose}{A boolean value.  If \code{TRUE}, a verbose information will
be given along iterations for tracing the convergence.  The default
value is \code{FALSE}.}
}
\description{
This function fits regular Cox proportional hazard model (Cox, 1972) by the
monotonic quadratic approximation algorithm (Bohning and Lindsay, 1988).  It
is not intended to replace \code{\link[survival]{coxph}} or
\code{\link[survival]{coxph.fit}} for fitting regular Cox model.  However,
it allows non-integer event indicator (between 0 and 1) when computing the
partial likelihood (Cox, 1975) and the profiled baseline hazard function,
which can be useful for some research problems where event indicators
contain uncertainty.
}
\references{
Cox, D. R. (1972). Regression Models and Life-Tables. Journal of the Royal
Statistical Society. Series B (Methodological), 34(2), 187--220.

Cox, D. R. (1975). Partial Likelihood. Biometrika, 62(2), 269--276.

B\"ohning, Dankmar, & Lindsay, B. G. (1988). Monotonicity of Quadratic-
Approximation Algorithms. Annals of the Institute of Statistical
Mathematics, 40(4), 641--663.
}
