% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCoxph.R
\name{iCoxph}
\alias{iCoxph}
\title{Integrative Cox Model for Uncertain Survival Data}
\usage{
iCoxph(formula, data, subset, na.action, contrasts = NULL,
       start = list(), control = list(), ...)
}
\arguments{
\item{formula}{\code{Survi} object specifying the covariates and response
variable in the model, such as \code{Survi(ID, time, event) ~ x1 + x2}.}

\item{data}{An optional data frame, list, or environment that contains the
covariates and response variables included in the model. If not found in
data, the variables are taken from \code{environment(formula)}, usually
the environment from which function \code{\link{iCoxph}} is called.}

\item{subset}{An optional logical vector specifying a subset of observations
to be used in the fitting process.}

\item{na.action}{An optional function that indicates what should the
procedure do if the data contains \code{NA}s.  The default is set by the
na.action setting of \code{\link[base]{options}}.  The "factory-fresh"
default is \code{\link[stats]{na.omit}}.  Other possible values inlcude
\code{\link[stats]{na.fail}}, \code{\link[stats]{na.exclude}}, and
\code{\link[stats]{na.pass}}.  \code{help(na.fail)} for details.}

\item{contrasts}{An optional list, whose entries are values (numeric
matrices or character strings naming functions) to be used as
replacement values for the contrasts replacement function and whose
names are the names of columns of data containing factors.  See
\code{contrasts.arg} of \code{\link[stats]{model.matrix.default}} for
details.}

\item{start}{An optional list of starting values for the parameters to be
estimated in the model.  See more in Section details.}

\item{control}{An optional list of parameters to control the maximization
process.  See more in Section details.}

\item{...}{Other arguments for future usage.}
}
\value{
A \code{\link{iCoxph-class}} object, whose slots include
\itemize{
    \item \code{call}: Function call.
    \item \code{formula}: Formula used in the model fitting.
    \item \code{nObs}: Number of observation.
    \item \code{data}: A processed data frame used for model fitting.
    \item \code{estimates}:
        \itemize{
            \item \code{beta}: Coefficient estimates.
            \item \code{pi}: Estimated parameters in prior multinomial
                distribution indicating the probabilities of corresponding
                record being true.
            \item \code{h0}: Estimated baseline hazard function.
        }
    \item \code{control}: The control list specified for model fitting.
    \item \code{start}: The initial guess specified for the parameters
        to be estimated.
    \item \code{na.action}: The procedure specified to deal with
        missing values in the covariate.
    \item \code{xlevels}: A list that records the levels in
        each factor variable.
    \item \code{contrasts}: Contrasts specified and used for each
        factor variable.
    \item \code{convergCode}: \code{code} returned by function
        \code{\link[stats]{nlm}}, which is an integer indicating why the
        optimization process terminated. \code{help(nlm)} for details.
    \item \code{partLogL}: Partial log-likelihood function under complete
        data for covariate coefficients.
    \item \code{logL}: Log-likelihood under observed data after convergence.
    \item \code{fisher}: Observed Fisher information matrix under complete
        data after convergence.
}
}
\description{
The function fits the integrative Cox model proposed by Wang (2017+) for
uncertain survival data due to imperfect data integration. The maximization
of the observed-data likelihood function follows expectation conditional
maximization (ECM) algorithm proposed by Meng (1993).
}
\details{
The argument \code{start} is an optional list that allows users to specify
starting value of each parameter in the model. The valid named elements are
given as follows:
\itemize{

\item \code{beta}: A numeric vector for starting values of coefficient
    estimates. The default values are the coefficient estimates from the
    regular Cox model only fitting on records without uncertainty.  If
    censoring rate among subjects having unique certain records is extremely
    high (> 0.99) or \code{\link[survival]{coxph}} returns any warning, such
    as perfect seperation problem between event indicator and one predictor,
    the starting values will be all zeros by default.

\item \code{censorRate}: A positive numeric vector taking values between 0
    and 1. If a vector of length more than one is given, multiple starting
    values on prior probabilities are applied. In each initialization,
    individual value in the vector is taken as parameter in the multinomial
    distribution indicating the prior probability of censoring record being
    true.  The remaining probability mass is equally assigned to the
    possible event records. For example, when \code{censorRate = 1} is
    specified, the extended model reduces to the regular Cox model fitting
    data before matching, which takes all the censoring records as true. On
    the contrary, when \code{censorRate = 0} is specified, the model does
    not take possible censoring records into account.  If \code{multiStart =
    FALSE}, the default value of \code{censorRate} is the sample censoring
    rate among subjects having records without uncertainty.

\item \code{piVec}: A numeric vector specifying the prior probability of
    each original record being true. The length of the vector has to be the
    same with number of rows of the data input. The values should be between
    0 and 1. By default, \code{censorRate} is used to generate appropriate
    prior probabilities.

\item \code{multiStart}: A logical value specifying whether multiple
    starting values of \code{censorRate} is needed. The argument is ignored
    when explicit \code{censorRate} or \code{piVec} is specified. The
    default value is \code{FALSE} for less computation time. However,
    enabling multiple starting values \code{multiStart = TRUE} is suggested
    to reduce the influence of possible local maximizer. If \code{TRUE},
    \code{censorRate} takes a grid from 0 to 1 with step 0.02 by default.

}

The argument \code{control} is an optional list that allows users to control
the process of ECM algorithm and each conditional maximization (CM) step.
The valid named elements are given as follows:
\itemize{

\item \code{gradtol}: A positive scalar giving the tolerance at which the
    scaled gradient is considered close enough to zero to terminate each CM
    step. The default value is \code{1e-6}.

\item \code{stepmax}: A positive scalar that gives the maximum allowable
    scaled step length for each CM step.  The default value is \code{1e2}.

\item \code{steptol}: A positive scalar providing the minimum allowable
    relative step length for each CM step.  The default value is
    \code{1e-6}.

\item \code{iterlim}: A positive integer specifying the maximum number of
    iterations to be performed before each CM is terminated. The default
    value is \code{1e2}.

\item \code{steptol_ECM}: A positive scalar that specifies the maximum
    allowable relative step size between each ECM iteration.  The default
    value is \code{1e-4}.

\item \code{iterlim_ECM}: A positive integer specifying the maximum number
    of iterations to be performed before the ECM algorithm is
    terminated. The default value is \code{1e2}.

}
Internally, the first four named elements are passed to function
\code{\link[stats]{nlm}}.
}
\examples{

## FIXME: add function generating simulation data
## and example simulated dataset for demonstration
## library(intsurv)
## iCoxph(Survi())

}
\references{
Wang, W., Aseltine, R., Chen, K., & Yan, J. (2017+).  Integrative Survival
Analysis with Uncertain Event Records from Imperfect Data Integration with
Application in a Suicide Risk Study. (working in progress)

Meng, X. & Rubin, D. (1993). Maximum Likelihood Estimation via the ECM
Algorithm: A General Framework. \emph{Biometrika}, 80(2), 267--278.
}
\seealso{
\code{\link{summary,iCoxph-method}} for summary of fitted model;
\code{\link{coef,iCoxph-method}} for estimated covariate coefficients;
\code{\link{bootSe}} for SE estimates from bootstrap method.
}
