% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsurv.R
\docType{package}
\name{intsurv-package}
\alias{intsurv-package}
\title{Integrative Survival Modeling}
\description{
Integrative Survival Modeling
}
\references{
Wang, W., Aseltine, R., Chen, K., & Yan, J. (2019).  Integrative Survival
Analysis with Uncertain Event Times in Application to a Suicide Risk
Study. \emph{Annals of Applied Statistics}. (in press)

Wang, W., Chen, K., Luo, C., & Yan, J. (2019+). Cox Cure Model with
Uncertain Event Status with application to a Suicide Risk
Study. \emph{Working in Progress}.
}
