% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsurv.R
\docType{package}
\name{intsurv-package}
\alias{intsurv-package}
\title{Integrative Survival Modeling}
\description{
The package \pkg{intsurv} provides implementations of
}
\details{
\itemize{

\item integrative Cox model with uncertain event times (Wang et al., 2020)

\item Cox cure rate model with uncertain event status (Wang et al., 2020)

}

It also contains other survival analysis routines, including regular Cox
cure rate model, regularized Cox cure rate model with elastic net penalty,
and weighted concordance index.
}
\references{
Wang, W., Aseltine, R. H., Chen, K., & Yan, J. (2020). Integrative Survival
Analysis with Uncertain Event Times in Application to A Suicide Risk
Study. \emph{Annals of Applied Statistics}, 14(1), 51--73.

Wang, W., Luo, C., Aseltine, R. H., Wang, F., Yan, J., & Chen, K. (2020).
Suicide Risk Modeling with Uncertain Diagnostic Records. \emph{arXiv
preprint arXiv:2009.02597}.
}
