% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsurv.R
\docType{package}
\name{intsurv-package}
\alias{intsurv-package}
\title{Integrative Survival Modeling}
\description{
The package \pkg{intsurv} provides implementations of
}
\details{
\itemize{

\item integrative Cox model with uncertain event times (Wang et al., 2019)

\item Cox cure rate model with uncertain event status (Wang et al., 2019+)

}

It also contains other survival analysis routines, including regular Cox
cure rate model, regularized Cox cure rate model with elastic net penalty,
and weighted concordance index.
}
\references{
Wang, W., Aseltine, R., Chen, K., & Yan, J. (2019).  Integrative Survival
Analysis with Uncertain Event Times in Application to a Suicide Risk
Study. \emph{Annals of Applied Statistics}. (in press)

Wang, W., Chen, K., Luo, C., & Yan, J. (2019+). Cox Cure Model with
Uncertain Event Status with application to a Suicide Risk
Study. \emph{Working in Progress}.
}
