% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show-method}
\alias{show-method}
\alias{show,coxphx-method}
\alias{show,coxphx.summary-method}
\title{Show Methods}
\usage{
\S4method{show}{coxphx}(object)

\S4method{show}{coxphx.summary}(object)
}
\arguments{
\item{object}{An object used to dispatch a method.}
}
\value{
object input (invisibly).
}
\description{
S4 class methods that display or summarize certain objects.
}
\details{
\itemize{

\item For \code{\link{coxphx-class}} object, it prints out a brief summary
    of the fitted model.

\item For \code{\link{coxphx.summary-class}} object, it prints out summary
      of a fitted model.

}
}
