% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Survi-class}
\alias{Survi-class}
\title{An S4 Class Representing Formula Response}
\description{
An S4 Class Representing Formula Response
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A numeric matrix object.}

\item{\code{ID}}{Identificator of each subject.}

\item{\code{time}}{Observerd event time or censoring time.}

\item{\code{event}}{The event indicator, 0 = censored, 1 = event.}
}}

