% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survi.R
\name{Survi}
\alias{Survi}
\title{Formula Response for Survival Data With Uncertain Event Times}
\usage{
Survi(ID, time, event, check = TRUE, ...)
}
\arguments{
\item{ID}{Identificator of each subject.}

\item{time}{Time of reccurence event or censoring.}

\item{event}{The status indicator, 0 = censored, 1 = event.}

\item{check}{A logical value specifying whether to perform check on input
data.}

\item{...}{Other arguments for future usage.  A warning will be thrown if
any invalid argument is specified.}
}
\value{
\code{Survi} object.  See \code{\link{Survi-class}} for details.
}
\description{
\code{Survi} returns an S4 class that represents formula response for
survival data with uncertain records due to imperfect data integration.  The
last letter 'i' in \code{Survi} represents 'integration'.
}
\examples{
## See examples of function 'iCoxph'
}
